Move Harvest specific functions to harvestUtils
